requirements.txt
# Python version: 3.10-3.12 (testado e recomendado)
# Para Windows, recomenda-se Python 3.12

# Core UI
customtkinter==5.2.2
Pillow>=10.0.0,<11.0.0

# Database
sqlalchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0,<3.0.0
supabase>=2.0.0,<3.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<3.0.0

# Authentication
bcrypt>=4.0.0,<5.0.0
PyJWT>=2.8.0,<3.0.0

# Date/Time
python-dateutil>=2.8.0,<3.0.0
tkcalendar>=1.6.1,<2.0.0


# Excel and PDF
openpyxl>=3.1.0,<4.0.0
reportlab>=4.0.0,<5.0.0

# SVG support for logos
cairosvg>=2.7.0,<3.0.0

# Windows-specific: uncomment if needed
# pywin32>=306  # Para integrações Windows

